requirements.txt
beautifulsoup4==4.9.3
Brotli==1.0.9
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
chart-studio==1.1.0
chromedriver-install==1.0.3
click==7.1.2
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.8.0
Flask-SeaSurf==0.2.2
future==0.18.2
geckodriver-autoinstaller==0.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
plotly==4.12.0
PyYAML==5.3.1
requests==2.25.0
requests-file==1.5.1
retrying==1.3.3
selenium==3.141.0
selenium-firefox==0.5.27
six==1.15.0
soupsieve==2.0.1
style==1.1.0
tldextract==3.1.0
ua-parser==0.10.0
update==0.0.1
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
youtube-dl==2020.12.5
